- Corrección de nomenclatura y reorganización de dto
