--- backend/src/encuestas/dto/create-opcion.dto.ts
@@ -1,11 +1,0 @@
-import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
-
-export class CreateOpcionDTO {
-  @IsString()
-  @IsNotEmpty()
-  texto: string;
-
-  @IsNumber()
-  @IsNotEmpty()
-  numero: number;
-}
