--- backend/src/encuestas/dto/create-pregunta.dto.ts
@@ -1,32 +1,0 @@
-import {
-  IsArray,
-  IsEnum,
-  IsNotEmpty,
-  IsNumber,
-  IsOptional,
-  IsString,
-  ValidateNested,
-} from 'class-validator';
-import { TiposRespuestaEnum } from '../../preguntas/enums/tipos-respuestas.enum';
-import { Type } from 'class-transformer';
-import { CreateOpcionDTO } from './create-opcion.dto';
-
-export class CreatePreguntaDTO {
-  @IsNumber()
-  @IsNotEmpty()
-  numero: number;
-
-  @IsString()
-  @IsNotEmpty()
-  texto: string;
-
-  @IsEnum(TiposRespuestaEnum)
-  @IsNotEmpty()
-  tipo: TiposRespuestaEnum;
-
-  @IsArray()
-  @IsOptional()
-  @ValidateNested({ each: true })
-  @Type(() => CreateOpcionDTO)
-  opciones?: CreateOpcionDTO[];
-}
