+++ backend/src/encuestas/services/encuestas.service.ts
@@ -4,7 +4,7 @@ import { Encuesta } from '../entities/encuesta.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { v4 } from 'uuid';
-import { CodigoTipoEnum } from '../enums/token-tipo.enum';
+import { TokenTipoEnum } from '../enums/token-tipo.enum';
 
 @Injectable()
 export class EncuestasService {
@@ -20,23 +20,23 @@ export class EncuestasService {
   }> {
     const encuesta: Encuesta = this.encuestasRepository.create({
       ...dto,
-      codigo_respuesta: v4(),
-      codigo_resultados: v4(),
+      token_respuesta: v4(),
+      token_resultados: v4(),
     });
 
     const encuestaGuardada = await this.encuestasRepository.save(encuesta);
 
     return {
       id: encuestaGuardada.id,
-      codigoRespuesta: encuestaGuardada.codigo_respuesta,
-      codigoResultados: encuestaGuardada.codigo_resultados,
+      codigoRespuesta: encuestaGuardada.token_respuesta,
+      codigoResultados: encuestaGuardada.token_resultados,
     };
   }
 
   async obtenerEncuesta(
     id: number,
     codigo: string,
-    codigoTipo: CodigoTipoEnum.RESPUESTA | CodigoTipoEnum.RESULTADOS,
+    codigoTipo: TokenTipoEnum.RESPUESTA | TokenTipoEnum.RESULTADOS,
   ): Promise<Encuesta> {
     const query = this.encuestasRepository
       .createQueryBuilder('encuesta')
@@ -45,11 +45,11 @@ export class EncuestasService {
       .where('encuesta.id = :id', { id });
 
     switch (codigoTipo) {
-      case CodigoTipoEnum.RESPUESTA:
+      case TokenTipoEnum.RESPUESTA:
         query.andWhere('encuesta.codigoRespuesta = :codigo', { codigo });
         break;
 
-      case CodigoTipoEnum.RESULTADOS:
+      case TokenTipoEnum.RESULTADOS:
         query.andWhere('encuesta.codigoResultados = :codigo', { codigo });
         break;
     }
