+++ backend/src/opciones/dto/create-opcione.dto.ts
@@ -1,1 +1,11 @@
-export class CreateOpcioneDto {}
+import { IsNotEmpty, IsNumber, IsString } from 'class-validator';
+
+export class CreateOpcioneDto {
+  @IsString()
+  @IsNotEmpty()
+  texto: string;
+
+  @IsNumber()
+  @IsNotEmpty()
+  numero: number;
+}
